✨ feat: DB 컬럼 변경에 따른 폼 수정
